requirements.txt
python-dotenv==1.0.0
pyTelegramBotAPI==4.10.0
aiohttp==3.8.4